requirements.txt
argparse==1.4.0
netaddr==0.8.0
ipaddr==2.2.0
requests==2.25.1
click==8.0.1
paramiko==2.7.2
texttable==1.6.3
PySocks==1.7.1
win_inet_pton==1.1.0
pyOpenSSL==20.0.1
flask==2.0.1
lockfile==0.12.2
bs4==0.0.1
pyspf==2.0.14
SQLAlchemy>=1.3.0
canari==3.3.10
py3DNS==3.2.1
censys==2.0.3
shodan==1.25.0
pycurl==7.43.0.6
xmltodict==0.12.0
rule==0.1.1
mechanize==0.4.5
scapy==2.4.5
psycopg2==2.9.1